DDQN + Checkpointing, Testing, and Plotting functionality
Used DDQN Network
Added checkpointing, testing, metric tracking, and plotting functionality 

When making a 'random' action, the solver makes a deterministic move 10% of the time. For example, if a human could solve the cell the environment is currently at, the 'random' action is the correct value of that cell. If a human can't solve, the 'random' action is to move away, which gets a positive reward.